requirements.txt
asgiref==3.7.2
click==8.1.7
colorama==0.4.6
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
install==1.3.5
numpy==1.26.1
Pillow==10.1.0
PyJWT==2.8.0
pytz==2023.3.post1
setuptools==68.2.2
sqlparse==0.4.4
tzdata==2023.3
